Refactor css modules into a single module to be used for future product sliders
