--- components/modules/Shoe/Shoe.module.css
@@ -1,47 +1,0 @@
-.root {
-  @apply relative grid items-start gap-1 grid-cols-1 overflow-x-hidden;
-  min-height: auto;
-}
-
-.main {
-  @apply relative px-0 pb-0 box-border  flex flex-col col-span-1;
-  min-height: 500px;
-}
-
-.sidebar {
-  @apply flex flex-col col-span-1 mx-auto max-w-7xl p-8 w-full h-full;
-}
-
-.sliderContainer {
-  @apply flex items-center justify-center overflow-x-hidden;
-}
-
-.imageContainer {
-  @apply text-center h-full relative;
-}
-
-.imageContainer > span {
-  height: 100% !important;
-}
-
-.sliderContainer .img {
-  @apply w-full h-full max-h-full object-cover;
-}
-
-@screen lg {
-  .root {
-    @apply grid-cols-12;
-  }
-
-  .main {
-    @apply mx-0 col-span-7;
-  }
-
-  .sidebar {
-    @apply col-span-5 py-6;
-  }
-
-  .imageContainer {
-    max-height: 600px;
-  }
-}
